extract default resolver
