+++ lib/rets.rb
@@ -266,3 +266,4 @@ require 'rets/parser/multipart'
 require 'rets/measuring_http_client'
 require 'rets/locking_http_client'
 require 'rets/client_progress_reporter'
+require 'rets/default_resolver'
